% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_funcs.R
\name{add_db_directory}
\alias{add_db_directory}
\title{Add DB Directory}
\usage{
add_db_directory(
  project_name,
  base_path = "/Shared/AML/small_dbs",
  sources = c("ccae", "mdcr", "medicaid"),
  years = 1:21,
  medicaid_years = 14:18,
  param_path = "/Shared/AML/params"
)
}
\arguments{
\item{project_name}{a name for the project. This should also correspond to a
disease group in the codeBuildr package}

\item{base_path}{a path of where to build the db}

\item{sources}{the sources to pull data from}

\item{years}{the years to pull data for}

\item{medicaid_years}{the years of medicaid data to pull from}

\item{param_path}{the path to the parameter folder}
}
\description{
Adds a directory for a small DB and creates an entry in the db build status
log
}
