% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helper_funcs.R
\name{get_dx_dates}
\alias{get_dx_dates}
\title{Get Diagnosis Dates for List of ICD-9 and ICD-10 diagnoses}
\usage{
get_dx_dates(setting, source, year, dx9_list, dx10_list, con, collect_n = 10)
}
\arguments{
\item{setting}{"inpatient", "outpatient", "rx"}

\item{source}{"ccae" or "mdcr" or c("ccae","mdcr")}

\item{year}{vector of years to collect. Note: if no argument is provided
all years will be selected}

\item{dx9_list}{vector of ICD-9 diagnoses to pull visits for}

\item{dx10_list}{vector of ICD-10 diagnoses to pull visits for}

\item{con}{a database connection}

\item{collect_n}{number of rows to collect}
}
\value{
A list containing three tibbles (timemap, outpatient keys, and inpatient keys)
}
\description{
This function is a wrapper around `get_dx9_dates()` and `get_dx10_dates()`
}
