% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helper_funcs.R
\name{get_table_vars}
\alias{get_table_vars}
\title{Get variables from a specified table}
\usage{
get_table_vars(table_name, source, year, vars, db_con, collect_n = 10)
}
\arguments{
\item{table_name}{the name of the table}

\item{source}{the source (ccae or mdcr)}

\item{year}{the year to collect from}

\item{vars}{the variables to collect}

\item{db_con}{a database connection}

\item{collect_n}{the number of variables to collect}
}
\description{
Get variables from a specified table
}
