% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_funcs.R
\name{pull_dx_dates}
\alias{pull_dx_dates}
\title{Pull diagnosis dates}
\usage{
pull_dx_dates(
  dx_codes_9,
  dx_codes_10,
  db_path,
  years,
  medicaid_years,
  cluster_size = 20,
  num_to_collect = 10
)
}
\arguments{
\item{dx_codes_9}{icd-9 diagnosis codes to collect}

\item{dx_codes_10}{icd-10 diagnosis codes to collect}

\item{db_path}{a string with the path to the databases folder}

\item{years}{years of data to collect for}

\item{medicaid_years}{years of medicaid data to collect for}

\item{cluster_size}{size of cluster to create for pulling data}

\item{num_to_collect}{number of rows to collect}
}
\value{
A list containing three tibbles (timemap, outpatient keys, and inpatient keys)
}
\description{
This function goes to the main database and pulls...
}
