% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_funcs.R
\name{pull_rx_encounters}
\alias{pull_rx_encounters}
\title{Pull rx encounters across multiple years}
\usage{
pull_rx_encounters(
  years,
  medicaid_years,
  ndc_codes,
  vars = c("enrolid", "ndcnum", "svcdate"),
  num_to_collect = 10,
  cluster_size = 20,
  db_path = "/Shared/Statepi_Marketscan/databases/Truven/"
)
}
\arguments{
\item{years}{vector of years to collect. Note: if no argument is provided
all years will be selected}

\item{medicaid_years}{years of medicaid data to collect for}

\item{ndc_codes}{ndc codes to pull}

\item{vars}{vector of variables to collect}

\item{num_to_collect}{the number of variables to collect}

\item{cluster_size}{size of cluster to create for pulling data}

\item{db_path}{path to Truven databases}
}
\description{
This function pulls columns from a specified table
}
