% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_funcs.R
\name{pull_tables}
\alias{pull_tables}
\title{Pull tables across multiple years}
\usage{
pull_tables()
}
\arguments{
\item{table_name}{name of table to pull}

\item{setting}{"inpatient", "outpatient", "rx"}

\item{source}{"ccae" or "mdcr" or c("ccae","mdcr")}

\item{year}{vector of years to collect. Note: if no argument is provided
all years will be selected}

\item{vars}{vector of variables to collect}

\item{num_to_collect}{the number of variables to collect}
}
\description{
This function pulls columns from a specified table
}
