% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_funcs.R
\name{update_build_params}
\alias{update_build_params}
\title{Update param}
\usage{
update_build_params(
  project_name,
  base_path = "/Shared/AML/small_dbs",
  sources = c("ccae", "mdcr", "medicaid"),
  years = 1:21,
  medicaid_years = 14:18,
  param_path = "/Shared/AML/params"
)
}
\arguments{
\item{project_name}{a name for the project. This should also correspond to a
disease group in the codeBuildr package}

\item{base_path}{a path of where to build the db}

\item{sources}{the sources to pull data from}

\item{years}{the years to pull data for}

\item{medicaid_years}{the years of medicaid data to pull from}

\item{param_path}{the path to the parameter folder}
}
\description{
Updates the status of the small_db_paths file
}
