% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_funcs.R
\name{update_build_status}
\alias{update_build_status}
\title{Update Build Status}
\usage{
update_build_status(
  project_name,
  description,
  param_path = "/Shared/AML/params"
)
}
\arguments{
\item{project_name}{a name for the project. This should also correspond to a
disease group in the codeBuildr package}

\item{description}{a description to update the status to}

\item{param_path}{the path to the parameter folder}
}
\description{
Updates the status of the small_db_paths file
}
