% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_funcs.R
\name{pull_dx_dates}
\alias{pull_dx_dates}
\title{Pull diagnosis dates}
\usage{
pull_dx_dates(
  dx_codes_9,
  dx_codes_10,
  db_path,
  cluster_size = 20,
  num_to_collect = 10
)
}
\arguments{
\item{setting}{"inpatient", "outpatient", "rx"}

\item{source}{"ccae" or "mdcr" or c("ccae","mdcr")}

\item{year}{vector of years to collect. Note: if no argument is provided
all years will be selected}

\item{dx_list}{vector of diagnoses to pull visits for}

\item{con}{a database connection}

\item{collect_n}{number of rows to collect}
}
\value{
A list containing three tibbles (timemap, outpatient keys, and inpatient keys)
}
\description{
This function goes to the main database and pulls...
}
